+++ spirv_cross/src/spirv.rs
@@ -347,45 +347,45 @@ pub enum Type {
     Unknown,
     Void,
     Boolean {
-        vecsize: usize,
+        vecsize: u32,
         array: Vec<u32>,
     },
     Char {
-        vecsize: usize,
+        vecsize: u32,
         array: Vec<u32>,
     },
     Int {
-        vecsize: usize,
+        vecsize: u32,
         array: Vec<u32>,
     },
     UInt {
-        vecsize: usize,
+        vecsize: u32,
         array: Vec<u32>,
     },
     Int64 {
-        vecsize: usize,
+        vecsize: u32,
         array: Vec<u32>,
     },
     UInt64 {
-        vecsize: usize,
+        vecsize: u32,
         array: Vec<u32>,
     },
     AtomicCounter {
         array: Vec<u32>,
     },
     Half {
-        vecsize: usize,
-        columns: usize,
+        vecsize: u32,
+        columns: u32,
         array: Vec<u32>,
     },
     Float {
-        vecsize: usize,
-        columns: usize,
+        vecsize: u32,
+        columns: u32,
         array: Vec<u32>,
     },
     Double {
-        vecsize: usize,
-        columns: usize,
+        vecsize: u32,
+        columns: u32,
         array: Vec<u32>,
     },
     Struct {
@@ -402,19 +402,19 @@ pub enum Type {
         array: Vec<u32>,
     },
     SByte {
-        vecsize: usize,
+        vecsize: u32,
         array: Vec<u32>,
     },
     UByte {
-        vecsize: usize,
+        vecsize: u32,
         array: Vec<u32>,
     },
     Short {
-        vecsize: usize,
+        vecsize: u32,
         array: Vec<u32>,
     },
     UShort {
-        vecsize: usize,
+        vecsize: u32,
         array: Vec<u32>,
     },
     ControlPointArray,
